build(teaser): Windows build/test patches for native compilation

+++ teaser/src/registration.cc
@@ -61,7 +61,7 @@ void teaser::ScalarTLSEstimator::estimate(const Eigen::RowVectorXd& X,
   double sum_xi = 0;
   double sum_xi_square = 0;
 
-  for (size_t i = 0; i < nr_centers; ++i) {
+  for (int i = 0; i < static_cast<int>(nr_centers); ++i) {
 
     int idx = int(std::abs(h.at(i).second)) - 1; // Indices starting at 1
     int epsilon = (h.at(i).second > 0) ? 1 : -1;
@@ -168,7 +168,7 @@ void teaser::ScalarTLSEstimator::estimate_tiled(const Eigen::RowVectorXd& X,
 #pragma omp parallel for shared(                                                                   \
     jh_bound, ih_bound, ranges_inverse_sum_vec, dot_X_weights_vec, dot_weights_consensus_vec,      \
     X_consensus_table, h_centers, weights, N, X, x_hat, x_cost, s, ranges, inner_loop_f)
-  for (size_t ih = 0; ih < ih_bound; ih += s) {
+  for (int ih = 0; ih < static_cast<int>(ih_bound); ih += static_cast<int>(s)) {
     for (size_t jh = 0; jh < jh_bound; jh += s) {
       for (size_t il = 0; il < s; ++il) {
         size_t i = ih + il;
@@ -183,7 +183,7 @@ void teaser::ScalarTLSEstimator::estimate_tiled(const Eigen::RowVectorXd& X,
     shared(jh_bound, ih_bound, ranges_inverse_sum_vec, dot_X_weights_vec,                          \
            dot_weights_consensus_vec, X_consensus_table, h_centers, weights, N, X, x_hat, x_cost,  \
            s, ranges, nr_centers, inner_loop_f)
-  for (size_t i = 0; i < nr_centers; ++i) {
+  for (int i = 0; i < static_cast<int>(nr_centers); ++i) {
     inner_loop_f(i, 0, jh_bound, N);
   }
 
@@ -192,7 +192,7 @@ void teaser::ScalarTLSEstimator::estimate_tiled(const Eigen::RowVectorXd& X,
     shared(jh_bound, ih_bound, ranges_inverse_sum_vec, dot_X_weights_vec,                          \
            dot_weights_consensus_vec, X_consensus_table, h_centers, weights, N, X, x_hat, x_cost,  \
            s, ranges, nr_centers, inner_loop_f)
-  for (size_t i = ih_bound; i < nr_centers; ++i) {
+  for (int i = static_cast<int>(ih_bound); i < static_cast<int>(nr_centers); ++i) {
     inner_loop_f(i, 0, 0, N);
   }
 
@@ -524,7 +524,7 @@ teaser::RobustRegistrationSolver::computeTIMs(const Eigen::Matrix<double, 3, Eig
   map->resize(2, N * (N - 1) / 2);
 
 #pragma omp parallel for shared(N, v, vtilde, map)
-  for (size_t i = 0; i < N - 1; i++) {
+  for (int i = 0; i < static_cast<int>(N) - 1; i++) {
     // Calculate some important indices
     // For each measurement, we compute the TIMs between itself and all the measurements after it.
     // For example:
